requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.0
pinyin==0.4.0
alibabacloud-nls-python-sdk==1.0.2 